% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHistoricCp.R
\name{addHistoricCp}
\alias{addHistoricCp}
\title{Add Historic CP Series to Balanced Accounts}
\usage{
addHistoricCp(historicCp, baCp, linkPeriod)
}
\arguments{
\item{historicCp}{data.frame with Period and Value columns for historic GFCF in CP.}

\item{baCp}{data.frame of balanced accounts in CP with Period and Value columns.}

\item{linkPeriod}{String denoting the Period at which to join historic and
balanced accounts (balanced accounts will start from the linkPeriod).}
}
\value{
data.frame with Period and Value columns with periods covering min(historicCp)
to max(baCp).
}
\description{
Add Historic CP Series to balanced accounts series at the defined reference period.
}
