% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHistoricCvm.R
\name{addHistoricCvm}
\alias{addHistoricCvm}
\title{Add Historic CVM Series to Balanced Accounts}
\usage{
addHistoricCvm(historicCp, historicDeflator, baCp, baCvm, linkPeriod)
}
\arguments{
\item{historicCp}{data.frame with Period and Value columns for historic GFCF in CP.}

\item{historicDeflator}{data.frame with Period and Value columns for historic
deflator. Typically as created by extractHistoricDeflators.}

\item{baCp}{data.frame of balanced accounts in CP with Period and Value columns.}

\item{baCvm}{data.frame of balanced accounts in CVM with Period and Value columns.}

\item{linkPeriod}{String denoting the Period at which to join historic and
balanced accounts (balanced accounts will start from the linkPeriod).}
}
\value{
data.frame with Period and Value columns with periods covering min(historicCp)
to max(baCvm).
}
\description{
Calculate Historic CVM Series from constant price data and historic deflators
and add to balanced accounts series.
}
\details{
Historic GFCF series in CP are converted to CVM using historic deflators.
Balanced accounts CVM series are appended to create a dataset spanning the
entire time period.

The historic CVM series is calculated by using an implied deflator from
balanced accounts data (CP / CVM at the link period). The historic
deflator is then referenced to this implied deflator:
referencedDeflator@period = (deflator@period / deflator@link) / impliedDeflator@link

The referenced deflator is then applied to the historic GFCF series to
convert it to CVM. Balanced accounts CVM values are appended from the link
period onwards.

Occasionally the balanced accounts CP or CVM values at the link period
are zero so unsuitable for creating an implied deflator. In these cases the
next non-zero period is used and a warning is issued.
}
