% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLifeLengths.R
\name{addLifeLengths}
\alias{addLifeLengths}
\title{Add life lengths.}
\usage{
addLifeLengths(gfcf, lifeLengthsPath)
}
\arguments{
\item{gfcf}{data.frame of GFCF series.}

\item{lifeLengthsPath}{string denoting path to life lengths Excel file.}
}
\value{
data.frame of gfcf with additional Min, Max, Average, and CoV columns.
}
\description{
Join Life Length Min, Max, Average, and CoVs to GFCF dataset.
}
\details{
The life lengths Excel file must include four sheets labelled AverageLifeLengths,
CoVs, Min, and Max. Each sheet must contain Sector, Industry, Asset columns and
additional columns for each period.
}
