% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOtherParams.R
\name{addOtherParams}
\alias{addOtherParams}
\title{Add Other Parameters to GFCF data.}
\usage{
addOtherParams(gfcf, paramsFile)
}
\arguments{
\item{gfcf}{data.frame of GFCF series containing Sector, Industry, Asset, Period, and Value columns.}

\item{paramsFile}{string denoting the path to the parameters CSV file containing Period, TaxAdjustmentFactor, UTIL, and CPI columns.}
}
\value{
data.frame with gfcf with a new "PriceIndex" column containing the deflators.
}
\description{
Add other parameters (TaxAdjustmentFactor, UTIL, CPI) to GFCF data.
}
\details{
The other parameters are read from a CSV file, checked for coverage, and then left-joined to the GFCF data.frame using the Period column.
}
