% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPriceIndex.R
\name{addPriceIndex}
\alias{addPriceIndex}
\title{Add PriceIndex to GFCF data.}
\usage{
addPriceIndex(gfcf, deflators)
}
\arguments{
\item{gfcf}{data.frame of GFCF series containing Sector, Industry, Asset, Period, and Value columns.}

\item{deflators}{data.frame containing Sector, Industry, Asset, Period, and Value columns.}
}
\value{
data.frame with gfcf with a new "PriceIndex" column containing the deflators.
}
\description{
Add PriceIndex to a prepared GFCF data.frame.
}
\details{
Deflators are checked for coverage and then left-joined to the GFCF data.frame.
}
