% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySplitSpec.R
\name{applySplits}
\alias{applySplits}
\title{Internal: Split a category into smaller categories and pro-rate its values}
\usage{
applySplits(.data, splits, existingCategory, newCategory, tol = 0.01)
}
\arguments{
\item{.data}{data.frame of GFCF data.}

\item{splits}{data.frame with one row per proportion, see details.}

\item{existingCategory}{String denoting the variable to split in .data. Common
to both \code{.data} and \code{splits}, e.g. "Sector".}

\item{newCategory}{String denoting the variable containing the more detailed
level of \code{existingCategory} in \code{splits}, e.g. "Sector_new". Note,
no spaces are allowed.}
}
\value{
data.frame with matching rows split out as per the proportions.
}
\description{
Splits a category (i.e. within a Sector, Industry, or Asset) into smaller
sub-categories and divides its values between the sub-categories using the
supplied proportions.
}
\details{
\code{.data} is typically a GFCF dataset, e.g. as imported by
extractBalAccountsCord, containing \code{Sector}, \code{Industry}, \code{Asset},
\code{Prices}, \code{Period}, and \code{Value}.

\code{splits} is a data.frame defining split proportions with one row per split,
containing a \code{Sector}, \code{Industry}, \code{Asset}, \code{Period}, and
\code{Proportion}, with an additional column denoting the smaller categories
e.g. "Sector_new".

Splits are matched to the data by matching on all standard columns:
\code{Sector}, \code{Industry}, \code{Asset}, \code{Period}.
The existing category is then replaced by \code{newCategory} and the
proportions are applied.

Only one category can be split at a time for a particular splits definition,
e.g. a splits definition can be used to split Sectors into smaller Sectors
or Assets into smaller Assets, but not both at once.
}
