% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySplitSpec.R
\name{applySplitSpec}
\alias{applySplitSpec}
\title{Apply a split specification to data and apportion its values.}
\usage{
applySplitSpec(.data, spec, existingCategory, newCategory, joinKeys, tol)
}
\arguments{
\item{.data}{data.frame of GFCF data to be split.}

\item{spec}{String path to a CSV file containing split specification or a data.frame in the same format.}

\item{existingCategory}{String denoting the variable to split in .data. Common
to both \code{.data} and \code{splits}, e.g. "Sector".}

\item{newCategory}{String denoting the variable in the split specification containing the more detailed
level of \code{existingCategory}, e.g. "Split". Note, no spaces are allowed.}

\item{joinKeys}{character vector of dimensions defining matching priority, e.g. c("Asset", "Industry", "Sector").
See Split Matching for details.}
}
\value{
data.frame with matching rows split out as per the proportions.
}
\description{
Split a category (a Sector, Industry, or Asset) into smaller
sub-categories and divide its values between the sub-categories using the
provided proportions.
}
\section{Split Specification File}{

The splits spec CSV file should have Sector, Industry, Asset, Split columns,
and additional columns for each Period containing the split proportions.
E.g. for a single sector split definition:

\tabular{llllrrr}{
Sector \tab Industry \tab Asset \tab Split \tab Y1990Q1 \tab Y1990Q2 \tab ...\cr
S.1PT \tab ALL \tab ALL \tab S.11PR \tab 0.8 \tab 0.8 \tab ...\cr
S.1PT \tab ALL \tab ALL \tab S.14   \tab 0.1 \tab 0.1 \tab ...\cr
S.1PT \tab ALL \tab ALL \tab S.15   \tab 0.1 \tab 0.1 \tab ...
}

A data.frame in the same format can be provided instead of a path to allow for
creation of split specifications on-the-fly.

Split proportions should be in the range 0-1 and should sum to 1 when grouped
by Sector, Industry, and Asset. Any splits with proportions of zero are removed.

The \emph{ALL} keyword can be used instead of specific Sectors, Industries, or
Assets in the specification file. Where \emph{ALL} occurs it will be expanded
to match the coverage in the \code{.data}.
}

\section{Split Matching}{


Splits are matched to the data by matching on all standard columns:
\code{Sector}, \code{Industry}, \code{Asset}, \code{Period}. Any data that
does not have a matching entry in the split specification is passed through
unchanged.

Splits will first be matched to the coverage using the most detailed
specification, where all dimensions of joinKeys match, i.e. by
Sector, Industry, and Asset. Any remaining data will be matched more generally
by removing a joinKey, e.g. by matching on only Sector and Industry. Finally,
The splits will be matched by just Sector.

The order of these iterations is controlled by the order of \code{joinKeys},
with each iteration removing the last element of \code{joinKeys} to make the
join more general.
}

