% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCoverage}
\alias{checkCoverage}
\title{Check Coverage}
\usage{
checkCoverage(timeseries1, timeseries2, dimensions)
}
\arguments{
\item{timeseries1}{data.frame in tidy format, typically containing Period, Value, and other categories such as Sector, Industry}

\item{timeseries2}{data.frame in tidy format, typically containing Period, Value, and other categories such as Sector, Industry}

\item{dimensions}{The common column names between the two time series.}
}
\description{
Check two time series dataframes for coverage, based on the specified dimensions
}
