% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkDimCoverage}
\alias{checkDimCoverage}
\title{Check Coverage of single dimension}
\usage{
checkDimCoverage(df1, df2, dimension)
}
\arguments{
\item{timeseries1}{data.frame in tidy format, typically containing Period, Value, and other categories such as Sector, Industry}

\item{timeseries2}{data.frame in tidy format, typically containing Period, Value, and other categories such as Sector, Industry}

\item{dimensions}{String of the common column name.}
}
\value{
printout of differences in categores between dataframes.
}
\description{
Check two time series dataframes for coverage, based on a single shared column and report back discrepancies
}
