% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandConfigSpec.R
\name{expandConfigSpec}
\alias{expandConfigSpec}
\title{Expand Configuration Specification}
\usage{
expandConfigSpec(spec, toCover, joinKeys)
}
\arguments{
\item{spec}{data.frame with one row per config specification, usually containing
Sectory, Industry, and Asset columns, plus colums of named arguments to pimir::pimConfig.}

\item{toCover}{data.frame with every unique combination of Sector, Industry, and Asset required.}

\item{joinKeys}{character vector of dimensions defining matching priority, e.g. c("Asset", "Industry", "Sector").}
}
\value{
data.frame with one row for each Sector/Industry/Asset combination in \code{toCover}.
}
\description{
Expands a configuration specification to match the provided coverage.
}
\details{
The configs will first be matched to the coverage using the most detailed
specification, where all joinKeys match, e.g. by Asset, Industry, and Sector
Then configs will be matched more generally by removing a joinKey, e.g. by
matching on only Asset and Industry. Finally, The configs will be matched
by just Asset.

The ordering of the matching will always go from most specific to least, but
which dimensions take precedence is controlled by the ordering of the joinKeys
character vector.

The first element of joinKeys should usually be Asset.

\code{toCover} is deduplicated before matching so for convenience a GFCF file
can be provided to create configs matching its coverage. Any configs defined
in the specification that are unmatched in the coverage are ignored.
}
