% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandSpec.R
\name{expandSpec}
\alias{expandSpec}
\title{Expand specification dataframe}
\usage{
expandSpec(spec, toCover, joinKeys)
}
\arguments{
\item{spec}{data.frame containing Sector, Industry, Asset columns and others}

\item{toCover}{data.frame containing Sector, Industry, Asset columns denoting the required coverage}

\item{joinKeys}{vector denoting the order of precedence for expanding Sector, Industry, Asset, e.g. c("Sector", "Industry", "Asset"). See Details.}
}
\value{
the spec data.frame with any "ALL" values expanded to meet the provided coverage.
}
\description{
Expand a specification dataframe to match a given coverage. Sector, Industry,
or Asset values containing the "ALL" wildcard will be expanded to match the
coverage.
}
\details{
Up to two of Sector/Industry/Asset can be "ALL" per spec row (but not all three).

Splits will first be matched to the coverage using the most detailed
specification, where all joinKeys match, e.g. by Sector, Industry, and Asset.
Then splits will be matched more generally by removing a joinKey, e.g. by
matching on only Sector and Industry. Finally, The splits will be matched
by just Sector.

The ordering of the matching will always go from most detailed to least, but
which dimensions take precedence is controlled by the ordering of the joinKeys
character vector.

The first element of joinKeys should be the main category that is being split,
i.e. if we are expanding a sector splits file the first element should be "Sector",
and should not contain the wildcard "ALL".
}
