% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySplitSpec.R
\name{expandSplitSpec}
\alias{expandSplitSpec}
\title{Expand Splits Specification}
\usage{
expandSplitSpec(spec, toCover, joinKeys, tidy = TRUE)
}
\arguments{
\item{spec}{data.frame with one row per split, containing Sector, Industry, and Asset columns.}

\item{toCover}{data.frame with every unique combination of Sector, Industry, and Asset required.}

\item{joinKeys}{character vector of dimensions defining matching priority, e.g. c("Sector", "Industry", "Asset").}

\item{tidy}{boolean denoting whether to tidy the resulting data.frame.
This will gather the period columns into one called Period, i.e. change the result from wide to long.}
}
\value{
data.frame with one row per split for each Sector/Industry/Asset and Period if tidy = TRUE.
}
\description{
Expands a splits specification to match the coverage in the data provided.
}
\details{
Splits will first be matched to the coverage using the most detailed
specification, where all joinKeys match, e.g. by Sector, Industry, and Asset.
Then splits will be matched more generally by removing a joinKey, e.g. by
matching on only Sector and Industry. Finally, The splits will be matched
by just Sector.

The ordering of the matching will always go from most detailed to least, but
which dimensions take precedence is controlled by the ordering of the joinKeys
character vector.

The first element of joinKeys should be the main category that is being split,
i.e. if we are expanding a sector splits file the first element should be "Sector".

\code{toCover} is deduplicated before matching so for convenience a GFCF file
could be provided to create splits matching its coverage. Any splits defined
in the specification that are unmatched in the coverage are ignored.

Any splits with proportions of zero are removed.
}
