% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractAdjustments.R
\name{extractAdjustments}
\alias{extractAdjustments}
\title{Extract Adjustments.}
\usage{
extractAdjustments(path)
}
\arguments{
\item{path}{String path to CSV file.}
}
\value{
Verified data.frame with Sector, Industry, Asset, Period, and K values.
}
\description{
Extract K-value adjustments from CSV file.
}
\details{
The CSV file should contain Sector, Industry, Asset, Period, K1CP, K3CP, K4CP,
K5CP, K61CP, and K62CP columns. Other columns may be included but will be ignored.

Each row must be unique with respect to Sector, Industry, Asset, and Period.
}
