% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBalAccountsCord.R
\name{extractBalAccountsCord}
\alias{extractBalAccountsCord}
\title{Extract balanced accounts from CORD outputs}
\usage{
extractBalAccountsCord(path, sheet = 1, skip = 3)
}
\arguments{
\item{path}{File path of the input spreadsheet.}

\item{sheet}{Sheet number within the Excel file. Defaults to first sheet.}

\item{skip}{Number of rows to skip before reading any data. Defaults to 3
based on typical CORD export.}
}
\value{
A tidy data.frame of net balanced accounts with \code{Sector},
 \code{Industry}, \code{Asset}, \code{Period}, and \code{Value} variables.
}
\description{
Extracts balanced accounts data from a standard CORD Excel export.
}
\details{
This function depends on the output format from CORD being consistent.

  Input files should have columns Sector, Industry, Asset & Product, Basis,
  followed by columns for each Period, e.g. Y1997Q1, Y1997Q2, etc.

  1. Data is parsed from the first worksheet in the provided Excel file path

  2. Empty rows (containing all \code{NA}s or zeros) are removed.

  3. Net values are calculated (Acquisitions - Disposals)

  If net GFCF figures have been pre-calculated they should be provided with a
  Basis of "NET" instead of "ACQ" or "DIS". Input files must
  have a Basis of either "NET" or "ACQ"/"DIS" but not mixing both.

  To reduce logging verbosity use \code{flog.threshold(WARN, name = "prepim")}
  using WARN, ERROR, or FATAL. Defaults to INFO level.
}
