% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConfigSpec.R
\name{extractConfigSpec}
\alias{extractConfigSpec}
\title{Extract Configuration Specification}
\usage{
extractConfigSpec(path)
}
\arguments{
\item{path}{String path to a CSV file containing specification.}
}
\value{
verified data.frame of specifications.
}
\description{
Extract a configuration specification file and check its consistency.
}
\details{
The configuration spec is expected to have Sector, Industry, Asset
columns, with additional columns for each configuration option.

See \code{pimir::pimConfig} for the argument names and options. Columns must match
the argument names exactly.

Specifications can include the word "ALL" to create default configurations,
e.g. Sector:ALL, Industry:ALL, Asset:TRANSPORT would define a specification
for TRANSPORT for all Sectors and Industries.
}
