% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractHistoricDeflators.R
\name{extractHistoricDeflators}
\alias{extractHistoricDeflators}
\title{Extract historic deflators from reference file}
\usage{
extractHistoricDeflators(path)
}
\arguments{
\item{x}{File path of the input spreadsheet.}
}
\value{
tidy data.frame of deflators with \code{Sector},
 \code{Industry}, and \code{Asset} variables, plus the time series.
}
\description{
Extracts historic deflators pre-defined CSV file.
}
\details{
This function depends on the historic file being a consistent format.

  1. Data is parsed from the CSV file

  2. Empty rows (containing all \code{NA}s or zeros) are removed.

  To reduce logging verbosity use \code{flog.threshold(WARN, name = "prepim")}
  using WARN, ERROR, or FATAL. The default is INFO.
}
