% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLifeLength.R
\name{extractLifeLength}
\alias{extractLifeLength}
\title{Extract Life Length sheet from Excel.}
\usage{
extractLifeLength(llFile, toCover, varName, sheet = 1)
}
\arguments{
\item{llFile}{Excel file with one sheet per life length spec (e.g. "CoV").}

\item{toCover}{data.frame containing all the Sector/Industry/Assets that "ALL" will be expanded to.}

\item{varName}{string denoting the variable type present in the sheet, i.e. Average, CoV, Min, Max.}

\item{sheet}{the sheet name or number containing a life length type}
}
\value{
data.frame with Sector, Industry, Asset, Period, and varName columns.
}
\description{
Extract life length sheet from Excel and expand Sector/Industry/Assets of "ALL".
}
\details{
Industry codes are left-padded with zeros if required, e.g. "1" becomes "01".
}
