% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSplitSpec.R
\name{extractSplitSpec}
\alias{extractSplitSpec}
\title{Extract Split Specification}
\usage{
extractSplitSpec(path)
}
\arguments{
\item{path}{String path to a CSV file containing splits.}
}
\value{
verified data.frame of splits.
}
\description{
Extract a splits specification file and check its consistency.
}
\details{
This function is usually called from \code{applySplitSpec}, but can be used
directly if adjustments to the split specification are required before
applying the splits.

The specification is expected to have Sector, Industry, Asset, Split columns, with
additional columns for each Period, containing split proportions,
for example Sector, Industry, Asset, Split, Y1990, Y1991, Y1992...

Split proportions should be in the range 0-1 and should sum to 1 when grouped
by Sector, Industry, and Asset.
}
