% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatIndustryCodes}
\alias{formatIndustryCodes}
\title{Format Industry Codes}
\usage{
formatIndustryCodes(industryCodes)
}
\arguments{
\item{industryCodes}{vector of industry codes}
}
\value{
character vector with single digits padded with leading zeros
}
\description{
Formats a character vector of industry codes into a standard format.
}
\details{
Industry codes can lose their formatting in Excel, e.g. "01" becomes "1".
This function left-pads and single digit codes with "0" and ensures data type is character
}
