% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSeries}
\alias{getSeries}
\title{Get Series}
\usage{
getSeries(.data)
}
\arguments{
\item{.data}{data.frame in tidy format containing Sector, Industry, and Asset}
}
\description{
Get the unique combinations of Sector, Industry, and Asset from a data.frame
}
