% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{historicToQuarterly}
\alias{historicToQuarterly}
\title{Spline historic data from yearly to quarterly}
\usage{
historicToQuarterly(.data, method = "div4")
}
\arguments{
\item{.data}{data.frame as created by \code{extractHistoric}}

\item{method}{string denoting which method to use to spread yearly Values
across four periods.}
}
\value{
data.frame with Values spread from yearly Periods to Quarterly
}
\description{
Spreads out quantities from yearly amounts to quarterly amounts
}
\details{
The default method is simply to divide the yearly amount by four.
}
