% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeMissingCvmCoverage.R
\name{imputeMissingCvmCoverage}
\alias{imputeMissingCvmCoverage}
\title{Impute Missing CVM Coverage}
\usage{
imputeMissingCvmCoverage(requiredCoverage, cp, cvm)
}
\arguments{
\item{requiredCoverage}{data.frame with Sector, Industry, Asset variables for the required coverage.}

\item{cp}{data.frame of Current Prices, e.g. as produced by \code{extractBalAccountsCord}. Must contain Industry TOTALs for required Sector/Asset coverage.}

\item{cvm}{data.frame of CVM, e.g. as produced by \code{extractBalAccountsCord} which has missing series. Must contain Industry TOTALs for required Sector/Asset coverage.}
}
\value{
A tidy data.frame of imputed balanced accounts with \code{Sector},
 \code{Industry}, \code{Asset}, \code{Period}, and \code{Value} variables.
}
\description{
Imputes CVM series where present in CP balanced accounts but not
  in CVM for given Sector, Industry, and Assets using Industry TOTALs.
}
\details{
The cp and cvm data.frames must include series for Industry TOTALs.
A new series is created for each row of the required \code{coverage}
  using the following method:

  1. Calculate implied deflator using Industry TOTALs for required
    Sector/Asset combination (CP TOTAL / CVM TOTAL)

  2. Apply the deflator to the required Sector/Industry/Asset CP series.

  If the required Sector/Asset TOTALs are not available in the CVM data, a
  deflator is created as above but using S.11PR/Asset as a fallback.
}
