% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPriceIndex.R
\name{linkDeflators}
\alias{linkDeflators}
\title{Link historic and new deflators.}
\usage{
linkDeflators(defHist, defRecent, linkPeriod)
}
\arguments{
\item{defHist}{Historic deflators up to and including the link period.}

\item{defRecent}{Recent deflators from the link period onwards.}

\item{linkPeriod}{String denoting the link period.}
}
\value{
data.frame of deflators from min(defHist$Period) to max(defRecent$Period).
}
\description{
Combine series of historic and new deflators, linking them together at the specified Period.
}
\details{
Historic deflators are referenced to the new deflators at the link period.
Both datasets must include data for the link period and should include Sector,
Industry, Asset, Period, and Value columns.

The recent deflators are returned as-is and the historic deflators are linked

The returned data.frame has Sector, Industry, Asset, Period, and Value columns.
}
